user custom ApiWrapper to add custom methods
* move fetchHasMany() hack to apiWrapper
* move selectMany()'s options out from UserGroup
