+++ lib/api/user-groups.js
@@ -3,11 +3,41 @@
 const router = require('express').Router()
 const {BadRequestError} = require('jsonapi-errors/lib/errors')
 
+const { filter } = require('lodash')
+const pluralize = require('pluralize')
+const Serializer = require('jsonapi-serializer').Serializer
+const foreignKey = require('../sql-builder/foreign-key')
+
 const {debugApi} = require('../utils/debug')
 const userGroup = require('../models/user-group')
 
 const ApiWrapper = require('./api-wrapper')
-const apiWrappedUserGroup = new ApiWrapper(userGroup)
+
+class CustomApiWrapper extends ApiWrapper {
+  fetchHasMany (relationModelName) {
+    const pluralRelationName = pluralize(relationModelName)
+    const fkName = foreignKey(this.name)
+    const relationModel = this.model.registry.model(relationModelName)
+
+    const userOptions = (relationModelName === 'user') ? { where: {hide: false}, orderBy: 'name' } : {}
+
+    return Promise.all([ this.model.selectMany(), relationModel.selectMany(userOptions) ])
+    .then(([modelRows, relationRows]) => {
+      const dataSet = modelRows.map((row) => {
+        row[pluralRelationName] = filter(relationRows, [fkName, row.id])
+        return row
+      })
+
+      const options = { attributes: this.model.sqlBuilder.columnsNames.concat(pluralRelationName) }
+      options[pluralRelationName] = { ref: 'id', attributes: relationModel.sqlBuilder.columnsNames }
+      const serializer = new Serializer(this.name, options)
+
+      return serializer.serialize(dataSet)
+    })
+  }
+}
+
+const apiWrappedUserGroup = new CustomApiWrapper(userGroup)
 
 function readAll (req, res, next) {
   debugApi('user-group#readAll')
