+++ lib/models/user-group.js
@@ -3,32 +3,7 @@
 const registry = require('./registry')
 const BaseModel = require('./base-model')
 
-const {filter} = require('lodash')
-const pluralize = require('pluralize')
-const Serializer = require('jsonapi-serializer').Serializer
-const foreignKey = require('../sql-builder/foreign-key')
-
-class UserGroup extends BaseModel {
-  fetchHasMany (relationModelName) {
-    const pluralRelationName = pluralize(relationModelName)
-    const fkName = foreignKey(this.name)
-    const relationModel = registry.model(relationModelName)
-
-    return Promise.all([ this.selectMany(), relationModel.selectMany() ])
-      .then(([modelRows, relationRows]) => {
-        const dataSet = modelRows.map((row) => {
-          row[pluralRelationName] = filter(relationRows, [fkName, row.id])
-          return row
-        })
-
-        const options = { attributes: this.sqlBuilder.columnsNames.concat(pluralRelationName) }
-        options[pluralRelationName] = { ref: 'id', attributes: relationModel.sqlBuilder.columnsNames }
-        const serializer = new Serializer(this.name, options)
-
-        return serializer.serialize(dataSet)
-      })
-  }
-}
+class UserGroup extends BaseModel {}
 
 UserGroup.schemaObject = {
   tableName: 'sPepTree',
