+++ lib/models/user.js
@@ -11,17 +11,6 @@ const registry = require('./registry')
 const BaseModel = require('./base-model')
 
 class User extends BaseModel {
-  /**
-   * override it to provide own options
-   */
-  selectMany () {
-    const options = {
-      where: {hide: false},
-      orderBy: 'name'
-    }
-    return super.selectMany(options)
-  }
-
   passwordVerify (id, password) {
     debug(`${this.name}#passwordVerify`)
